backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosmtplib==5.1.0
aiosqlite==0.22.1
alphashape==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.10.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
attrs==26.1.0
bcrypt==4.1.3
beautifulsoup4==4.14.3
black==26.3.1
boto3==1.42.75
botocore==1.42.75
brotli==1.2.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.0.post2
charset-normalizer==3.4.6
click==8.3.1
click-log==0.4.0
Crawl4AI==0.8.6
cryptography==46.0.5
cssselect==1.4.0
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
emergentintegrations==0.1.0
fake-useragent==2.2.0
fastapi==0.115.6
fastuuid==0.14.0
filelock==3.25.2
flake8==7.3.0
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.7.2
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==9.0.0
iniconfig==2.3.0
isort==8.0.1
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jq==1.11.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.3.1
librt==0.8.1
limits==5.8.0
litellm==1.80.0
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.1
motor==3.6.0
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.3
oauthlib==3.3.1
openai==2.30.0
packaging==26.0
pandas==3.0.1
passlib==1.7.4
patchright==1.58.2
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
playwright==1.58.0
playwright-stealth==2.0.3
pluggy==1.6.0
postgrest==2.28.3
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
psycopg2-binary==2.9.11
pwdlib==0.3.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pyflakes==3.4.0
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pymongo==4.9.2
pyOpenSSL==26.0.0
pyparsing==3.3.2
PyPDF2==3.0.1
pyroaring==1.0.4
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
rank-bm25==0.2.2
realtime==2.28.3
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
resend==2.27.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
s3transfer==0.16.0
s5cmd==0.2.0
scipy==1.17.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8.3
starlette==0.41.3
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
trimesh==4.11.5
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
unclecode-litellm==1.81.13
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.34.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
